refactor: simplify chess component hierarchy and container layout
