+++ packages/plugins/plugin-chess/src/capabilities/react-surface.tsx
@@ -16,7 +16,7 @@ export default () =>
     createSurface({
       id: meta.id,
       role: ['article', 'section', 'card--intrinsic', 'card--extrinsic', 'card--popover', 'card--transclusion'],
-      // TODO(burdon): Could this be standardized so that we don't require a subject property (like below)?
+      // TODO(burdon): Could this be standardized so that we don't require a subject property.
       filter: (data): data is { subject: ChessType } => Obj.instanceOf(ChessType, data.subject),
       component: ({ data, role }) => <ChessContainer game={data.subject} role={role} />,
     }),
