+++ packages/ui/react-ui-gameboard/src/Gameboard/Gameboard.tsx
@@ -3,14 +3,14 @@
 //
 
 import { monitorForElements } from '@atlaskit/pragmatic-drag-and-drop/element/adapter';
-import React, { type CSSProperties, forwardRef, type PropsWithChildren, useCallback, useEffect, useState } from 'react';
+import React, { forwardRef, type PropsWithChildren, useCallback, useEffect, useState } from 'react';
 
 import { log } from '@dxos/log';
 import { type ThemedClassName } from '@dxos/react-ui';
-import { mx } from '@dxos/react-ui-theme';
 
 import { GameboardContext, type GameboardContextType } from './context';
 import { type GameboardModel, isLocation, isPiece, type Move, type PieceRecord } from './types';
+import { mx } from '@dxos/react-ui-theme';
 
 type GameboardRootProps = PropsWithChildren<{
   model?: GameboardModel;
@@ -76,20 +76,15 @@ const GameboardRoot = ({ children, model, onDrop }: GameboardRootProps) => {
 
 GameboardRoot.displayName = 'Gameboard.Root';
 
-type GameboardContentsProps = ThemedClassName<
-  PropsWithChildren<{
-    style?: CSSProperties;
-  }>
->;
+const containFragment = 'is-[min(100cqw,100cqh)] bs-[min(100cqw,100cqh)]';
 
-/**
- * Container centers the board.
- */
-const GameboardContent = forwardRef<HTMLDivElement, GameboardContentsProps>(
-  ({ children, classNames, style }, forwardedRef) => {
+type GameboardContentProps = ThemedClassName<PropsWithChildren<{ contain?: boolean }>>;
+
+const GameboardContent = forwardRef<HTMLDivElement, GameboardContentProps>(
+  ({ children, classNames, contain }, forwardedRef) => {
     return (
-      <div ref={forwardedRef} style={style} className='flex w-full h-full justify-center overflow-hidden'>
-        <div className={mx('max-w-full max-h-full content-center aspect-square', classNames)}>{children}</div>
+      <div ref={forwardedRef} role='none' className={mx(classNames)}>
+        {contain ? <div className={containFragment}>{children}</div> : children}
       </div>
     );
   },
@@ -100,4 +95,4 @@ export const Gameboard = {
   Content: GameboardContent,
 };
 
-export type { GameboardRootProps, GameboardContentsProps };
+export type { GameboardRootProps, GameboardContentProps };
